Move generate_exp tests to test_evaluation until we decide on home
